Allow additional commands after make
By supplying an additional command to the `TINKER` environment variable,
this command is appended to the list of commands executed within the
docker container.

Additionally, the image name is now quoted, to prevent escaping from the
docker command.

Lastly, the `-t` flag has been removed, because this makes the terminal
behave interactively when it shouldn't.