Do not remove docker volumes as dependency
The docker volumes can be automatically removed if the development setup
is run for the first time.

Since setting it up as dependency does not result in a new file, the
dependency will always run, which is unwanted behavior.

The dependency is merged back within the make target for the development
environment.