Allow interactive shell and TTY in `docker-make`
The `aws` command now uses an interactive shell and outputs to TTY.
Because the `docker-make` did not provide a TTY, it could not bind to
one and hence the verification of the `aws --version` command failed
within CI.

By making the shell interactive and using a TTY, the installation for
`docker-compose-development` started asking for user input. This is now
patched, as it provides defaults when INTERACTIVE mode is disabled.